Refactor Spring Security tests
