+++ web/src/test/java/ua/com/vertex/security/AuthenticationTest.java
@@ -1,0 +1,101 @@
+package ua.com.vertex.security;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.test.context.support.WithAnonymousUser;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import ua.com.vertex.context.TestConfig;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestBuilders.formLogin;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestBuilders.logout;
+import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.authenticated;
+import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.unauthenticated;
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(classes = TestConfig.class)
+@WebAppConfiguration
+@ActiveProfiles("test")
+public class AuthenticationTest {
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mockMvc;
+    private static final String CORRECT_PASSWORD = "123456";
+    private static final String EXISTING_EMAIL = "44@test.com";
+    private static final String INCORRECT = "incorrect";
+    private static final String ADMIN_EMAIL = "33@test.com";
+
+    @Before
+    public void setUp() {
+        mockMvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .apply(springSecurity())
+                .build();
+    }
+
+    @Test
+    @WithAnonymousUser
+    public void userTestFormLoginWithCorrectLoginAndPassword() throws Exception {
+        Collection<? extends GrantedAuthority> authorities = Arrays.asList(new SimpleGrantedAuthority("USER"));
+        mockMvc.perform(formLogin("/logIn").user(EXISTING_EMAIL).password(CORRECT_PASSWORD))
+                .andExpect(status().isOk())
+                .andExpect(forwardedUrl("/loggedIn"))
+                .andExpect(authenticated().withUsername(EXISTING_EMAIL))
+                .andExpect(authenticated().withAuthorities(authorities));
+    }
+
+    @Test
+    @WithAnonymousUser
+    public void adminTestFormLoginWithCorrectLoginAndPassword() throws Exception {
+        Collection<? extends GrantedAuthority> authorities = Arrays.asList(new SimpleGrantedAuthority("ADMIN"));
+        mockMvc.perform(formLogin("/logIn").user(ADMIN_EMAIL).password(CORRECT_PASSWORD))
+                .andExpect(status().isOk())
+                .andExpect(forwardedUrl("/loggedIn"))
+                .andExpect(authenticated().withUsername(ADMIN_EMAIL))
+                .andExpect(authenticated().withAuthorities(authorities));
+    }
+
+    @Test
+    @WithAnonymousUser
+    public void testFormLoginWithIncorrectPassword() throws Exception {
+        mockMvc.perform(formLogin("/logIn").user(EXISTING_EMAIL).password(INCORRECT))
+                .andExpect(status().isFound())
+                .andExpect(redirectedUrl("/logIn?error"))
+                .andExpect(unauthenticated());
+    }
+
+    @Test
+    @WithAnonymousUser
+    public void testFormLoginWithIncorrectLogin() throws Exception {
+        mockMvc.perform(formLogin("/logIn").user(INCORRECT).password(CORRECT_PASSWORD))
+                .andExpect(status().isFound())
+                .andExpect(redirectedUrl("/logIn?error"))
+                .andExpect(unauthenticated());
+    }
+
+    @Test
+    @WithMockUser
+    public void testFormLogout() throws Exception {
+        mockMvc.perform(logout("/logOut"))
+                .andExpect(status().isFound())
+                .andExpect(redirectedUrl("/"))
+                .andExpect(unauthenticated());
+    }
+}
